requirements.txt
# --- OFFICIAL EXCHANGE SDKs ---
binance-connector
python-kraken-sdk
coinbase-advanced-py

# --- DATA & MATH ---
pandas>=2.0.0
numpy>=1.24.0

# --- DATA FEEDS & ASYNC ---
websockets>=12.0
aiohttp>=3.8.0

# --- INFRASTRUCTURE & SECURITY ---
python-dotenv>=1.0.0
cryptography>=41.0.0
ecdsa
requests
certifi

# --- MONITORING & DASHBOARD ---
streamlit>=1.28.0
plotly>=5.17.0
psutil>=5.9.0
Flask>=1.0
apscheduler